[#6] restructuring api server dir
