+++ mas-api/api/api_initializer.py
@@ -2,8 +2,8 @@ import inject
 from utils.config import Config
 
 from api.auth.service.google_oauth2_service import GoogleOAuth2Service
-from api.database.database_connection_manager import DatabaseConnectionManager
-from api.database.mysql_connection_manager import MySQLConnectionManager
+from api.common.database.database_connection_manager import DatabaseConnectionManager
+from api.common.database.mysql_connection_manager import MySQLConnectionManager
 from api.user.repository.user_repository import UserRepository
 from api.user.service.user_service import UserService
