+++ mas-api/api/script/entity/script.py
@@ -2,7 +2,7 @@ from datetime import datetime
 
 from sqlalchemy.orm import Mapped, mapped_column
 
-from api.database.database_connection_manager import Base
+from api.common.database import Base
 from api.utils.datetime_utils import get_now_datetime_by_timezone
