+++ mas-api/api/user/repository/user_repository.py
@@ -1,6 +1,6 @@
 from sqlalchemy import delete, select, update
 
-from api.database.database_connection_manager import DatabaseConnectionManager
+from api.common.database import DatabaseConnectionManager
 from api.user.entity.user import User
 from api.user.exception.user_exception import UserNotFoundException
 from api.utils.datetime_utils import get_now_datetime_by_timezone
